CASMINST-3341 add kdump workaround script
This is the same workaround script from CASMINST-1072 with added automation to
run it on every NCN.

CASMINST-1072 was pulled from Shasta v1.4.2 since it was incomplete and
confusing to customers.  It was not added back in since it was fixed and
integrated with systemd in Shasta v1.5.

This workaround should be run on fresh installs and any time NCNs are rebuilt.

Signed-off-by: Jacob Salmela <jacob.salmela@hpe.com>